Convert ROBOT.md file to use json format
Using the json format will make it easier for us to store meta data to categorize and tag projects. The garden is filling up (which is AMAZING!!) but its starting to get a little cluttered. I am hoping to introduce more organization soon and switching the ROBOT.MD files to use json is the first step.